tests/test_patterns.py
"""
Test configuration and sample tests for Flight Sales Dashboard
This file demonstrates testing patterns and can be expanded with actual tests
"""

import pytest
import pandas as pd
from datetime import datetime, date
from unittest.mock import Mock, patch, MagicMock


class TestDatabaseConnections:
    """Test database connection functionality"""
    
    @patch('src.core.database.GraphDatabase.driver')
    @patch('src.core.database.MongoClient')
    def test_init_connections_success(self, mock_mongo, mock_neo4j):
        """Test successful database connection initialization"""
        from src.core.database import init_connections
        
        # Mock successful connections
        mock_neo4j.return_value.verify_connectivity.return_value = None
        mock_mongo.return_value.admin.command.return_value = None
        
        driver, mongo_client, mongo_db = init_connections()
        
        assert driver is not None
        assert mongo_client is not None
        assert mongo_db is not None
    
    @patch('src.core.database.GraphDatabase.driver')
    def test_init_connections_neo4j_failure(self, mock_neo4j):
        """Test Neo4j connection failure handling"""
        from src.core.database import init_connections
        
        mock_neo4j.side_effect = Exception("Neo4j connection failed")
        
        driver, mongo_client, mongo_db = init_connections()
        
        assert driver is None
        assert mongo_client is None
        assert mongo_db is None


class TestAnalytics:
    """Test analytics functions"""
    
    def test_get_sales_by_date_empty_result(self):
        """Test daily sales calculation with empty dataset"""
        from src.core.analytics import get_sales_by_date
        
        mock_collection = Mock()
        mock_collection.aggregate.return_value = []
        
        start_date = datetime(2023, 3, 10)
        end_date = datetime(2023, 3, 20)
        
        df, duration = get_sales_by_date(mock_collection, start_date, end_date)
        
        assert df.empty
        assert isinstance(duration, float)
        assert duration >= 0
    
    def test_get_sales_by_date_with_data(self):
        """Test daily sales calculation with sample data"""
        from src.core.analytics import get_sales_by_date
        
        mock_collection = Mock()
        mock_collection.aggregate.return_value = [
            {
                "_id": "2023-03-10",
                "daily_sales": 1000000,
                "daily_orders": 100
            },
            {
                "_id": "2023-03-11",
                "daily_sales": 1200000,
                "daily_orders": 120
            }
        ]
        
        start_date = datetime(2023, 3, 10)
        end_date = datetime(2023, 3, 11)
        
        df, duration = get_sales_by_date(mock_collection, start_date, end_date)
        
        assert len(df) == 2
        assert 'date' in df.columns
        assert 'daily_sales' in df.columns
        assert 'daily_orders' in df.columns


class TestDataValidation:
    """Test data validation and transformation"""
    
    def test_date_format_conversion(self):
        """Test date format conversion"""
        test_date_str = "2023-03-10"
        converted = pd.to_datetime(test_date_str)
        
        assert isinstance(converted, pd.Timestamp)
        assert converted.year == 2023
        assert converted.month == 3
        assert converted.day == 10
    
    def test_sales_aggregation(self):
        """Test sales data aggregation"""
        data = {
            'date': ['2023-03-10', '2023-03-10', '2023-03-11'],
            'sales': [100000, 200000, 150000]
        }
        df = pd.DataFrame(data)
        df['date'] = pd.to_datetime(df['date'])
        
        daily_sales = df.groupby('date')['sales'].sum()
        
        assert daily_sales.loc[pd.to_datetime('2023-03-10')] == 300000
        assert daily_sales.loc[pd.to_datetime('2023-03-11')] == 150000


class TestPerformance:
    """Test performance-related functionality"""
    
    def test_query_execution_time_tracking(self):
        """Test that execution times are properly tracked"""
        import time
        from src.core.analytics import get_sales_by_date
        
        mock_collection = Mock()
        mock_collection.aggregate.return_value = []
        
        start_date = datetime(2023, 3, 10)
        end_date = datetime(2023, 3, 20)
        
        df, duration = get_sales_by_date(mock_collection, start_date, end_date)
        
        # Duration should be non-negative
        assert duration >= 0
    
    def test_batch_vs_individual_queries(self):
        """Test that batch queries execute faster"""
        # This would require timing actual queries
        # Conceptual test showing the pattern
        
        batch_query_time = 0.05  # 50ms
        individual_queries_time = 0.50  # 500ms (10 queries)
        
        improvement = ((individual_queries_time - batch_query_time) / 
                      individual_queries_time) * 100
        
        assert improvement > 50  # At least 50% improvement expected


class TestBusinessLogic:
    """Test business logic and insights generation"""
    
    def test_average_order_value_calculation(self):
        """Test average order value calculation"""
        total_sales = 1000000
        total_orders = 100
        
        avg_order_value = total_sales / total_orders
        
        assert avg_order_value == 10000
    
    def test_performance_improvement_calculation(self):
        """Test performance improvement percentage calculation"""
        time_without_optimization = 1.0
        time_with_optimization = 0.4
        
        improvement = ((time_without_optimization - time_with_optimization) / 
                      time_without_optimization) * 100
        
        assert improvement == 60  # 60% improvement


class TestErrorHandling:
    """Test error handling and edge cases"""
    
    def test_division_by_zero_handling(self):
        """Test handling of division by zero"""
        total_sales = 100000
        total_orders = 0
        
        avg_order_value = total_sales / total_orders if total_orders > 0 else 0
        
        assert avg_order_value == 0
    
    def test_empty_dataframe_handling(self):
        """Test handling of empty DataFrames"""
        df = pd.DataFrame()
        
        if df.empty:
            result = "No data available"
        else:
            result = "Data processed"
        
        assert result == "No data available"
    
    def test_null_value_handling(self):
        """Test handling of null values in data"""
        data = {
            'origin': ['CGK', 'SIN', None],
            'sales': [100000, 200000, 150000]
        }
        df = pd.DataFrame(data)
        
        df_clean = df.dropna()
        
        assert len(df_clean) == 2


# Pytest configuration
@pytest.fixture
def sample_data():
    """Fixture providing sample test data"""
    return {
        'date': ['2023-03-10', '2023-03-11', '2023-03-12'],
        'sales': [100000, 120000, 110000],
        'orders': [10, 12, 11]
    }


@pytest.fixture
def mock_mongodb():
    """Fixture providing mock MongoDB connection"""
    return Mock()


@pytest.fixture
def mock_neo4j():
    """Fixture providing mock Neo4j connection"""
    return Mock()


# Running tests
# pytest tests/test_patterns.py -v
# pytest tests/test_patterns.py::TestAnalytics -v
# pytest tests/ --cov=. --cov-report=html
