requirements.txt
# Core Web Framework
streamlit>=1.28.0

# Data Processing and Analysis
pandas>=2.0.0

# Database Drivers
pymongo>=4.5.0
neo4j>=5.12.0

# Data Visualization
plotly>=5.15.0

# Environment Configuration
python-dotenv>=1.0.0
